much results saved

+++ Modules/Calibration/PinHole.cc
@@ -38,7 +38,7 @@ void PinHole::unproject(const Eigen::Vector2f& p2D, Eigen::Vector3f& p3D) {
     p3D[2] = 1.0;
 }
 
-void PinHole::unprojectWithZ(const Eigen::Vector2f& p2D, Eigen::Vector3f& p3D, float z) {
+void PinHole::unprojectWithZ(const Eigen::Vector2f& p2D, Eigen::Vector3f& p3D, double z) {
     p3D[0] = z * (p2D[0] - cx) / fx;
     p3D[1] = z * (p2D[1] - cy) / fy;
     p3D[2] = z;
